+++ CHANGELOG.md
@@ -141,6 +141,9 @@
   [Jumhyn](https://github.com/jumhyn)
   [#3295](https://github.com/realm/SwiftLint/issues/3295)
 
+* Fix finding a nested config when a single file path is passed.  
+  [Seth Friedman](https://github.com/sethfri)
+
 ## 0.40.2: Demo Unit
 
 #### Breaking